update  connection form
move to defaultUI
