--- ConnectForm.xaml
@@ -1,61 +1,0 @@
-﻿<Window x:Class="TheFlightSims.HyperVDPD.ConnectForm"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TheFlightSims.HyperVDPD"
-        Title="Connect to Computer" 
-        Height="200"
-        Width="400"
-        ResizeMode="NoResize" 
-        SizeToContent="WidthAndHeight">
-    <Grid Width="400">
-        <Grid.Margin>
-            <Thickness Bottom="5" Top="2"></Thickness>
-        </Grid.Margin>
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <Grid Grid.Row="0" Margin="10,0,10,0">
-            <Label VerticalAlignment="Center">Connect to another computer using domain or IP</Label>
-        </Grid>
-        <Grid Grid.Row="1" Margin="10,0,10,0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="100"/>
-                <ColumnDefinition Width="280"/>
-            </Grid.ColumnDefinitions>
-            <Label Grid.Column="0" VerticalAlignment="Center">Computer Name</Label>
-            <TextBox x:Name="ComputerName_TextBox" Grid.Column="1" VerticalAlignment="Center"/>
-        </Grid>
-        <Grid Grid.Row="2" Margin="10,0,10,0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="100"/>
-                <ColumnDefinition Width="280"/>
-            </Grid.ColumnDefinitions>
-            <Label Grid.Column="0" VerticalAlignment="Center">User Name</Label>
-            <TextBox x:Name="UserName_TextBox" Grid.Column="1" VerticalAlignment="Center"/>
-        </Grid>
-        <Grid Grid.Row="3" Margin="10,0,10,0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="100"/>
-                <ColumnDefinition Width="280"/>
-            </Grid.ColumnDefinitions>
-            <Label Grid.Column="0" VerticalAlignment="Center">Password</Label>
-            <PasswordBox x:Name="Password_TextBox" Grid.Column="1" VerticalAlignment="Center"/>
-        </Grid>
-        <Grid Grid.Row="4" Margin="10,0,10,0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="90"/>
-                <ColumnDefinition Width="200"/>
-                <ColumnDefinition Width="90"/>
-            </Grid.ColumnDefinitions>
-            <Button Margin="0,0,5,0" Grid.Column="0" VerticalAlignment="Center" Click="ConnetToExternal_Button">Connect</Button>
-            <Button Margin="5,0,5,0" Grid.Column="1" VerticalAlignment="Center" Click="ConnectToLocal_Button">Connect to Local Computer</Button>
-            <Button Margin="5,0,0,0" Grid.Column="2" VerticalAlignment="Center" Click="Cancel_Button">Cancel</Button>
-        </Grid>
-    </Grid>
-</Window>
