+++ DefaultUI/ConnectForm.xaml
@@ -1,0 +1,63 @@
+﻿<Window x:Class="TheFlightSims.HyperVDPD.DefaultUI.ConnectForm"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:TheFlightSims.HyperVDPD"
+        mc:Ignorable="d"
+        Title="Connect To Another Computer"
+        Height="180"
+        Width="400"
+        MinHeight="180"
+        MinWidth="400"
+        MaxHeight="180"
+        MaxWidth="400"
+        ResizeMode="NoResize"
+        SizeToContent="WidthAndHeight">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="3*"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <Grid Grid.Row="0">
+            <Grid.Margin>
+                <Thickness Left="5" Right="5"></Thickness>
+            </Grid.Margin>
+            <Label VerticalAlignment="Center">Connect to another computer using domain or IP</Label>
+        </Grid>
+        <Grid Grid.Row="1">
+            <Grid.Margin>
+                <Thickness Left="5" Right="5"></Thickness>
+            </Grid.Margin>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="2*"/>
+                <ColumnDefinition Width="5*"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition/>
+                <RowDefinition/>
+                <RowDefinition/>
+            </Grid.RowDefinitions>
+            <Label Grid.Row="0" Grid.Column="0" VerticalAlignment="Center">Computer Name</Label>
+            <TextBox Grid.Row="0" Grid.Column="1" x:Name="ComputerName_TextBox" VerticalAlignment="Center"/>
+            <Label Grid.Row="1" Grid.Column="0" VerticalAlignment="Center">User Name</Label>
+            <TextBox Grid.Row="1" Grid.Column="1" x:Name="UserName_TextBox" VerticalAlignment="Center"/>
+            <Label Grid.Row="2" Grid.Column="0" VerticalAlignment="Center">Password</Label>
+            <PasswordBox Grid.Row="2" Grid.Column="1" x:Name="Password_TextBox" VerticalAlignment="Center"/>
+        </Grid>
+        <Grid Grid.Row="2">
+            <Grid.Margin>
+                <Thickness Left="5" Right="5" Bottom="5"></Thickness>
+            </Grid.Margin>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="4*"/>
+                <ColumnDefinition Width="*"/>
+            </Grid.ColumnDefinitions>
+            <Button Margin="0,0,5,0" Grid.Column="0" VerticalAlignment="Center" Click="ConnetToExternal_Button">Connect</Button>
+            <Button Margin="5,0,5,0" Grid.Column="1" VerticalAlignment="Center" Click="ConnectToLocal_Button">Connect to Local Computer</Button>
+            <Button Margin="5,0,0,0" Grid.Column="2" VerticalAlignment="Center" Click="Cancel_Button">Cancel</Button>
+        </Grid>
+    </Grid>
+</Window>
